[xs] i18n fixes

+++ ckan/lib/base.py
@@ -188,20 +188,23 @@ class BaseController(WSGIController):
         # This also improves the cachability of our pages as cookies
         # prevent proxy servers from caching content unless they have
         # been configured to ignore them.
-        for cookie in request.cookies:
-            if cookie.startswith('ckan') and cookie not in ['ckan']:
-                response.delete_cookie(cookie)
-            # Remove the ckan session cookie if not used e.g. logged out
-            elif cookie == 'ckan' and not c.user and not h.are_there_flash_messages():
-                if session.id:
-                    if not session.get('lang'):
-                        session.delete()
-                else:
-                    response.delete_cookie(cookie)
-            # Remove auth_tkt repoze.who cookie if user not logged in.
-            elif cookie == 'auth_tkt' and not session.id:
-                response.delete_cookie(cookie)
 
+        # we need to be careful with the /user/set_lang/ URL as this
+        # creates a cookie.
+        if not environ.get('HTTP_PATH', '').startswith('/user/set_lang/'):
+            for cookie in request.cookies:
+                if cookie.startswith('ckan') and cookie not in ['ckan']:
+                    response.delete_cookie(cookie)
+                # Remove the ckan session cookie if not used e.g. logged out
+                elif cookie == 'ckan' and not c.user and not h.are_there_flash_messages():
+                    if session.id:
+                        if not session.get('lang'):
+                            session.delete()
+                    else:
+                        response.delete_cookie(cookie)
+                # Remove auth_tkt repoze.who cookie if user not logged in.
+                elif cookie == 'auth_tkt' and not session.id:
+                    response.delete_cookie(cookie)
         try:
             return WSGIController.__call__(self, environ, start_response)
         finally:
